+++ FINALTEST/src/Main.java
@@ -1,5 +1,5 @@
 import display.SubPanels;
-import tile.TileManager;
+import menu.*;
 
 import javax.swing.*;
 
@@ -16,7 +16,7 @@ public class Main {
 
     public Main() {
         mainFrame = new Frame(1600,900);
-        subPanels = new SubPanels();;
+        subPanels = new SubPanels();
         loginPanel = new LoginPanel(mainFrame, subPanels);
         menuPanel = new MenuPanel(mainFrame, subPanels);
         shopPanel = new ShopPanel(mainFrame);
